requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2018.8.24        # via python-telegram-bot, requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.3.1       # via python-telegram-bot
future==0.16.0            # via python-telegram-bot
idna==2.7                 # via cryptography, requests
pycparser==2.19           # via cffi
pysocks==1.6.8
python-telegram-bot==11.1.0
requests==2.19.1
six==1.11.0               # via cryptography
urllib3==1.23             # via requests
